requirements.txt
asgiref==3.8.1
Django==5.1.6
django-mdeditor==0.1.20
Markdown==3.7
pillow==11.1.0
python-decouple==3.8
sqlparse==0.5.3
